Backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pymupdf>=1.23.0
python-docx>=1.1.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
groq>=0.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
